Bean Scopes and Lifecycle Methods
The Singleton Scope means that the same instance of pizzaDeliveryService will always be used regardless of the number of bean requests. The same object (memory location) will always be used. In this example, thePizzaDeliveryService and theSecondPizzaDeliveryService are located in the same memory location.
The Prototype Scope means that a new instance of pizzaDeliveryService will be created each time for each bean request. In this example, thePizzaDeliveryService is in one memory location and theSecondPizzaDeliveryService is in another location.